refactor(socket): scope PR #714 to the socket component only
Per review, keep this PR to the socket layer. Revert the changes that belong to
a follow-up (after this + the cmake PR merge):
- components/rtps_embedded/.../EsppTransport.cpp: restore the original raw
  setsockopt SO_RCVBUF call (the migration to Socket::set_receive_buffer_size
  will follow separately).
- undo the socket_msvc.hpp -> socket_win32.hpp rename (it forced include changes
  in rtsp/ and lib/espp.hpp); keep the file as socket_msvc.hpp with the
  generalized _WIN32 guards. The rename can follow with the caller migration.

Net effect: PR #714 now touches only components/socket/ (the new set_option /
buffer-size API, _MSC_VER->_WIN32 guards, sock_type_t=SOCKET, narrowing guards,
WSAStartup call_once, getsockopt optlen type).
