+++ components/rtsp/include/rtsp_server.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "socket_win32.hpp"
+#include "socket_msvc.hpp"
 
 #include <chrono>
 #include <memory>
