+++ components/rtsp/include/rtsp_session.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "socket_win32.hpp"
+#include "socket_msvc.hpp"
 
 #include <functional>
 #include <memory>
