+++ components/socket/include/udp_socket.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "socket_win32.hpp"
+#include "socket_msvc.hpp"
 
 #include <optional>
 #include <span>
